Suppress VS2015 warning.

+++ tiny_gltf_loader.h
@@ -486,7 +486,7 @@ static inline bool is_base64(unsigned char c) {
 }
 
 std::string base64_decode(std::string const &encoded_string) {
-  int in_len = encoded_string.size();
+  int in_len = static_cast<int>(encoded_string.size());
   int i = 0;
   int j = 0;
   int in_ = 0;
